requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
pytorch-lightning==2.1.0
hydra-core==1.3.2
torcheval==0.0.7
Pillow==11.1.0
gdown==5.2.0
tensorboard==2.17.1
nibabel==5.3.2
matplotlib==3.10.0
numpy==1.26.4
wandb==0.19.2